Adding design Start Game Fragment.
