--- app/src/main/java/com/example/snakegame_app/ui/StartGame/StartGameFragment.kt
@@ -1,35 +1,0 @@
-package com.example.snakegame_app.ui.StartGame
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.snakegame_app.databinding.FragmentStartGameBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-
-class StartGameFragment : Fragment() {
-
-    private var _binding: FragmentStartGameBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentStartGameBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-    }
-
-}
