+++ app/src/main/res/anim/blink_fade_out.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha android:fromAlpha="0.0"
+        android:toAlpha="1.0"
+        android:interpolator="@android:anim/accelerate_interpolator"
+        android:duration="4000"
+        android:repeatMode="reverse"
+        android:repeatCount="infinite" />
+</set>
