+++ app/src/main/res/anim/fade_in.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha
+        android:duration="5000"
+        android:fromAlpha="0.1"
+        android:toAlpha="1.0" />
+</set>
