+++ app/src/main/res/anim/fade_out.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha
+        android:duration="5000"
+        android:fromAlpha="1.0"
+        android:toAlpha="0.0"
+        />
+
+</set>
