+++ app/src/main/res/anim/translate_snake_1.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate
+        android:duration="3000"
+        android:fromXDelta="0%"
+        android:toXDelta="450%"
+        android:repeatCount="infinite"/>
+</set>
