+++ app/src/main/res/layout/fragment_start_game.xml
@@ -6,19 +6,18 @@
     android:background="@color/green_600"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.StartGame.StartGameFragment">
+    tools:context=".ui.startgame.StartGameFragment">
 
     <TextView
         android:id="@+id/textTitle"
-        android:layout_width="@dimen/layout_size_0dp"
+        style="@style/TextViewTitle"
+        android:layout_width="@dimen/size_0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/margin_space_24dp"
-        android:layout_marginEnd="@dimen/margin_space_24dp"
-        android:layout_marginBottom="@dimen/margin_space_32dp"
+        android:layout_marginBottom="@dimen/margin_space_64dp"
         android:text="@string/title_game_name"
         app:layout_constraintBottom_toTopOf="@+id/guideline1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintEnd_toStartOf="@+id/imageApple1"
+        app:layout_constraintStart_toEndOf="@+id/imageApple2" />
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline1"
@@ -28,16 +27,79 @@
         app:layout_constraintGuide_percent="0.5" />
 
     <Button
-        android:id="@+id/button"
-        android:layout_width="@dimen/layout_size_0dp"
+        android:id="@+id/startButton"
+        android:layout_width="@dimen/size_0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/margin_space_24dp"
         android:layout_marginTop="@dimen/margin_space_16dp"
         android:layout_marginEnd="@dimen/margin_space_24dp"
+        style="@style/MainButton"
         android:text="@string/start_game_text"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="@+id/guideline1" />
+
+    <ImageView
+        android:id="@+id/imagePath2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="@dimen/margin_space_32dp"
+        android:src="@drawable/path"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <ImageView
+        android:id="@+id/imageApple1"
+        android:layout_width="@dimen/size_35dp"
+        android:layout_height="@dimen/size_50dp"
+        android:layout_marginEnd="@dimen/margin_space_24dp"
+        android:layout_marginBottom="@dimen/margin_space_64dp"
+        android:rotation="@dimen/rotate_grade_23"
+        android:src="@drawable/apple"
+        app:layout_constraintBottom_toTopOf="@+id/guideline1"
+        app:layout_constraintEnd_toEndOf="parent" />
+
+    <ImageView
+        android:id="@+id/imageApple2"
+        android:layout_width="@dimen/size_35dp"
+        android:layout_height="@dimen/size_50dp"
+        android:layout_marginStart="@dimen/margin_space_24dp"
+        android:layout_marginBottom="@dimen/margin_space_64dp"
+        android:rotation="@dimen/rotate_grade_negative_23"
+        android:src="@drawable/apple"
+        app:layout_constraintBottom_toTopOf="@+id/guideline1"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <ImageView
+        android:id="@+id/imageSnake2"
+        android:layout_width="@dimen/size_130dp"
+        android:layout_height="@dimen/size_50dp"
+        android:src="@drawable/snake"
+        app:layout_constraintBottom_toBottomOf="@+id/imagePath2"
+        app:layout_constraintStart_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/imagePath2" />
+
+
+    <ImageView
+        android:id="@+id/imagePath1"
+        android:layout_width="@dimen/size_0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/margin_space_32dp"
+        android:src="@drawable/path"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/imageSnake1"
+        android:layout_width="@dimen/size_130dp"
+        android:layout_height="@dimen/size_50dp"
+        android:rotation="@dimen/rotate_grade_180"
+        android:src="@drawable/snake"
+        app:layout_constraintBottom_toBottomOf="@+id/imagePath1"
+        app:layout_constraintEnd_toStartOf="@+id/imagePath1"
+        app:layout_constraintTop_toTopOf="@+id/imagePath1" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
