+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/startGameFragment"
-        android:name="com.example.snakegame_app.ui.StartGame.StartGameFragment"
+        android:name="com.example.snakegame_app.ui.startgame.StartGameFragment"
         android:label="fragment_start_game"
         tools:layout="@layout/fragment_start_game" />
 </navigation>
