+++ app/src/main/res/values/styles.xml
@@ -1,7 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
+    <!-- texts style -->
+    <style name="TextViewTitle">
+        <item name="android:textColor">@color/white</item>
+        <item name="android:textSize">@dimen/text_size_48sp</item>
+        <item name="android:textAlignment">center</item>
+        <item name="android:gravity">center</item>
+        <item name="android:textStyle">bold</item>
+    </style>
+    <!-- texts style -->
 
-
+    <!-- buttons style -->
+    <style name="MainButton">
+        <item name="android:textColor">@color/black</item>
+        <item name="android:inputType">textPersonName</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:textSize">@dimen/text_size_25sp</item>
+        <item name="android:background">@drawable/background_button</item>
+    </style>
+    <!-- buttons style -->
 
 </resources>
